Figured It Out!
Special thanks to Chris G and Sprinter for the insight!

http://stackoverflow.com/questions/36758375/iterating-over-a-2d-array-er
ror/36762287#36762287

When I was iterating over my 2D array I kept returning 0 for my sum of
ints instead of a negative value. maxSumValue was automatically set to
0 causing the problem.
Thanks for the helpful tip, I set the maxSumValue to -Integer.Max_Value
to set the value to the most negative value Java 8 has.

Big thanks to you both!